Step 8: Replace PhotoExecutionContext.builder with services + emitPluginEvent
PhotoExecutionContext now exposes:
- services: BuilderServices (replaces direct AfilmoryBuilder reference)
- emitPluginEvent: EmitPluginEventFn (function-shaped, no class dependency)

Photo pipeline call sites updated:
- photo/processor.ts: passes builder.services and a bound emitPluginEvent fn
- photo/image-pipeline.ts: reads via .services and .emitPluginEvent
- plugins/geocoding.ts: same migration at its runWithPhotoExecutionContext
  call sites

Side moves to keep contracts as pure leaves:
- PluginRunState moves to core/contracts/plugin-ref.ts (re-exported from
  plugins/manager.ts for back-compat)
- BuilderPluginEvent / BuilderPluginEventPayloads move to
  core/contracts/plugin-events.ts (re-exported from plugins/types.ts)

Circular deps: 11 -> 9. The logger-adapter <-> execution-context pair
is now broken; remaining cycles all involve plugins/types.ts and the
photo/processor <-> image-pipeline pair (Steps 9 and 11).
